refactor(finance/imports): extract types and utils to keep files under 200 lines
Move AiCounters, ProcessContext, TransactionProcessResult, createAiCounters
from process-transaction.ts into process-transaction-helpers.ts; re-export
from process-transaction.ts to preserve the public API.

Move computeProposals, buildChangeSet, RuleProposal type out of
RuleCreationStep.tsx into rule-creation/utils.ts. RuleCreationStep.tsx now
contains only the React component tree (148 lines).
