+++ apps/pops-api/src/modules/finance/imports/lib/process-transaction.ts
@@ -9,48 +9,20 @@ import {
   buildUncertainFromAi,
   buildUncertainNoMatch,
   type AiCategorizationResult,
+  type AiCounters,
+  type ProcessContext,
+  type TransactionProcessResult,
 } from './process-transaction-helpers.js';
 import { isTransferOrIncomeRow } from './transfer-classifier.js';
 
+export {
+  createAiCounters,
+  type AiCounters,
+  type ProcessContext,
+  type TransactionProcessResult,
+} from './process-transaction-helpers.js';
+
 import type { ParsedTransaction, ProcessedTransaction } from '../types.js';
-import type { AliasMap, EntityLookupMap } from './entity-matcher.js';
-
-export interface AiCounters {
-  aiError: AiCategorizationError | null;
-  aiFailureCount: number;
-  aiApiCalls: number;
-  aiCacheHits: number;
-  totalInputTokens: number;
-  totalOutputTokens: number;
-  totalCostUsd: number;
-}
-
-export interface ProcessContext {
-  entityLookup: EntityLookupMap;
-  aliases: AliasMap;
-  knownTags: string[];
-  importBatchId: string;
-}
-
-export interface TransactionProcessResult {
-  matched?: ProcessedTransaction;
-  uncertain?: ProcessedTransaction;
-  failed?: ProcessedTransaction;
-  batchStatus: 'success' | 'failed';
-  errorEntry?: { description: string; error: string };
-}
-
-export function createAiCounters(): AiCounters {
-  return {
-    aiError: null,
-    aiFailureCount: 0,
-    aiApiCalls: 0,
-    aiCacheHits: 0,
-    totalInputTokens: 0,
-    totalOutputTokens: 0,
-    totalCostUsd: 0,
-  };
-}
 
 function tryEntityMatch(
   transaction: ParsedTransaction,
